Adding Delivery Confirmation functionality to PackageServiceOptions for domestic shipments. Was told by UPS support that:
"Please try to request a Adult Signature under PackageServiceOptions and not ShipmentServiceOptions. Usually, ShipmentServiceOptions is for International packages. "

When originally using Delivery Confirmation on ShipmentServiceOptions, I was receiving the error, which is now resolved with the committed changes.

"The requested accessory option is unavailable between the selected locations."

An example of code to utilize this on a domestic shipment is:

```
    $deliveryConfirmation = new \Ups\Entity\DeliveryConfirmation;
    $deliveryConfirmation->setDcisType(2);

    $packageServiceOptions = new \Ups\Entity\PackageServiceOptions();
    $packageServiceOptions->setDeliveryConfirmation($deliveryConfirmation);
    $package->setPackageServiceOptions($packageServiceOptions);
```